backend/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn==0.24.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL适配器
alembic==1.12.1  # 数据库迁移工具

# 数据验证与模型
pydantic==2.4.2

# 环境变量
python-dotenv==1.0.0

# 安全
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1

# 其他工具
request==2.32.5
python-multipart==0.0.6